Add apollo server to the GateWay microservice (#438)
* Installed apollo server and graphql

* Initial apollo server setup.

* Add mock blog service in order to stitch the schema.

* graphql installation.

* Fix path and port.

* Add apollo http link

* Add blog to docker compose.

* Basic schema stitching.

* Fix annoying typescript errors.
The includes section tels the typescript language server which files this config is used on.
it path was wrong so typescript language server didn't find the correct files so whatever we did in this file didn't take effect.
Now that this points to the right files, what we change here will take effect.

* We don't use docker compose in production anymore.
So i've deleted docker-compose.prod.yml and merged all the yamls to one.

* Revert "We don't use docker compose in production anymore."
Well.... I forogot that we're using `prod` docker compose in travis.
which is ok. we should keep it. maybe rename it so it's not
confusing.

This reverts commit 78a20daca25b3f123962bada1b412bfc9e466e97.

* Move gateway port definition to dev compose.

* Add dist.

* Remove all unused code.

* Fix TS build

* Fix ts build (install types)

* Missing nodemon

* Keep node_modules folder so it won't get deleted by docker-compose volume.

* Added async operations util and use the retry function in remote schema creation.

* Ignore unreachable services.

* Remove unused import.

* Deal with case when no schemas found (no services online)

* New line

* Rename blog to cms.

* forgot to rename in package.json